requirements.txt
pandas >= 1.3.5
json >= 2.0.9
numpy >= 1.21.5
pickle >= 4.0
PyTorch == 1.9.0
Torchvision == 0.10.0
Torchtext == 0.10.0
opencv-python == 4.1.2
speechrecognition == 3.8.1
sentence-transformers == 2.2.0
deepface == 0.0.74
rdflib >= 6.1.1